requirements.txt
requests==2.31.0
PyGithub==2.1.1
python-frontmatter==1.0.0
